% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-AbsoluteAbundanceData.R
\docType{class}
\name{AbsoluteAbundanceData-class}
\alias{AbsoluteAbundanceData-class}
\alias{AbsoluteAbundanceData}
\title{Absolute Abundance Data}
\description{
A class for working with count data from a microbial or ecological assay.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data.frame of integer abundance counts with species as columns and samples as rows}

\item{\code{sampleMetadata}}{A data.frame of metadata about the samples with samples as rows and metadata variables as columns}

\item{\code{recordIdColumn}}{The name of the column containing IDs for the samples. All other columns will be treated as abundance values.}

\item{\code{ancestorIdColumns}}{A character vector of column names representing parent entities of the recordIdColumn.}

\item{\code{imputeZero}}{A logical indicating whether NA/ null values should be replaced with zeros.}
}}

