% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_alphaDiv.R
\name{alphaDiv}
\alias{alphaDiv}
\title{Alpha diversity}
\usage{
alphaDiv(
  otu,
  method = c("shannon", "simpson", "evenness"),
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{otu}{data.frame with samples as rows, taxa as columns}

\item{method}{string defining the the alpha diversity method. Accepted values are 'shannon','simpson', and 'evenness'}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
something that's useful. TBD
}
\description{
This function returns alpha diversity values for each sample.
}
