% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_alphaDiv.R
\name{alphaDivApp}
\alias{alphaDivApp}
\title{Alpha diversity}
\usage{
alphaDivApp(
  otu,
  methods = c("shannon", "simpson", "evenness"),
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{otu}{data.frame with samples as rows, taxa as columns}

\item{methods}{vector of strings defining the the beta diversity dissimilarity methods to use. Must be a subset of c('shannon','simpson','evenness').}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
something that's useful. TBD
}
\description{
This function returns json with alpha diversity values for each sample.
}
