% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_betaDiv.R
\name{betaDiv}
\alias{betaDiv}
\title{Beta diversity}
\usage{
betaDiv(
  otu,
  method = c("bray", "jaccard", "jsd"),
  k = 2,
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{otu}{data.frame with samples as rows, taxa as columns}

\item{method}{string defining the the beta diversity dissimilarity method. Accepted values are 'bray','jaccard', and 'jsd'}

\item{k}{integer determining the number of pcoa dimensions to return}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
who knows
}
\description{
This function returns pcoa coordinates calculated from the beta diversity dissimilarity matrix.
}
