% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_betaDiv.R
\name{betaDivApp}
\alias{betaDivApp}
\title{Beta diversity app}
\usage{
betaDivApp(
  otu,
  methods = c("bray", "jaccard", "jsd"),
  k = 2,
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{otu}{data.frame with samples as rows, taxa as columns}

\item{methods}{vector of strings defining the the beta diversity dissimilarity methods to use. Must be a subset of c('bray','jaccard','jsd').}

\item{k}{integer determining the number of pcoa dimensions to return}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
we'll see.
}
\description{
This function returns the name of a json file with beta diversity results.
}
