% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Comparator.R
\docType{class}
\name{Comparator-class}
\alias{Comparator-class}
\alias{Comparator}
\title{Comparator}
\description{
A class for representing a variable that will be used to compare samples between two groups. The variable's
values will be used to split samples into groups.
}
\section{Slots}{

\describe{
\item{\code{variable}}{A VariableMetadata}

\item{\code{groupA}}{BinList}

\item{\code{groupB}}{BinList}
}}

