% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Compute.R
\docType{class}
\name{ComputeResult-class}
\alias{ComputeResult-class}
\alias{ComputeResult}
\title{Compute Result}
\description{
A class for consistently representing the results of computes.
This includes their representation in R, as JSON and how they are written to files.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data.frame of values where computed variables are columns and samples rows.}

\item{\code{recordIdColumn}}{The name of the column containing IDs for the samples. All other columns will be treated as computed values.}

\item{\code{computedVariableMetadata}}{veupathUtils::VariableMetadataList detailing the computed variables.}

\item{\code{computationDetails}}{An optional message about the computed results.}

\item{\code{parameters}}{A record of the input parameters used to generate the computed results.}
}}

