% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-correlation.R
\name{correlation,AbundanceData,AbundanceData-method}
\alias{correlation,AbundanceData,AbundanceData-method}
\title{Correlation of two AbundanceData objects}
\usage{
\S4method{correlation}{AbundanceData,AbundanceData}(
  data1,
  data2,
  method = c("spearman", "pearson"),
  verbose = c(TRUE, FALSE),
  proportionNonZeroThreshold = 0.5,
  varianceThreshold = 0,
  stdDevThreshold = 0
)
}
\arguments{
\item{data1}{AbundanceData object.}

\item{data2}{AbundanceData object.}

\item{method}{string defining the type of correlation to run. The currently supported values are 'spearman' and 'pearson'}

\item{verbose}{boolean indicating if timed logging is desired}

\item{proportionNonZeroThreshold}{numeric threshold to filter features by proportion of non-zero values across samples}

\item{varianceThreshold}{numeric threshold to filter features by variance across samples}

\item{stdDevThreshold}{numeric threshold to filter features by standard deviation across samples}
}
\value{
ComputeResult object
}
\description{
This function returns the correlation of all columns of an AbundanceData object with appropriate columns of a second AbundanceData object.
}
