% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-correlation.R
\name{correlation,AbundanceData,missing-method}
\alias{correlation,AbundanceData,missing-method}
\title{Correlation of abundance data and metadata}
\usage{
\S4method{correlation}{AbundanceData,missing}(
  data1,
  data2,
  method = c("spearman", "pearson"),
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{data1}{AbundanceData object. Will correlate abundance variables with specified variables in data2}

\item{method}{string defining the type of correlation to run. The currently supported values are 'spearman' and 'pearson'}

\item{verbose}{boolean indicating if timed logging is desired}
}
\description{
This function returns the correlation of all columns of an AbundanceData object with appropriate columns of a SampleMetadata object.
}
