% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-correlation.R
\name{correlation,AbundanceData,SampleMetadata-method}
\alias{correlation,AbundanceData,SampleMetadata-method}
\title{This is where the names should go, because they know what's data1 and data 2 and can do the appropriate renaming}
\usage{
\S4method{correlation}{AbundanceData,SampleMetadata}(
  data1,
  data2,
  method = c("spearman", "pearson"),
  verbose = c(TRUE, FALSE),
  variables = NULL,
  data1Name = "data1",
  data2Name = "data2"
)
}
\description{
This is where the names should go, because they know what's data1 and data 2 and can do the appropriate renaming
}
