% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-correlation.R
\name{correlation,data.table,data.table-method}
\alias{correlation,data.table,data.table-method}
\title{Correlation}
\usage{
\S4method{correlation}{data.table,data.table}(
  data1,
  data2,
  method = c("spearman", "pearson"),
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{data1}{data.table with columns as variables. All columns must be numeric. One row per sample.}

\item{data2}{data.table with columns as variables. All columns must be numeric. One row per sample. Will correlate all columns of data2 with all columns of data1.}

\item{method}{string defining the type of correlation to run. The currently supported values are 'spearman' and 'pearson'}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
data.frame with correlation coefficients
}
\description{
This function returns correlation coefficients for all columns in one data table with all columns in a second data table.
}
