% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-correlation.R
\name{correlation,data.table,missing-method}
\alias{correlation,data.table,missing-method}
\title{Correlation}
\usage{
\S4method{correlation}{data.table,missing}(
  data1,
  data2,
  method = c("spearman", "pearson", "sparcc"),
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{data1}{data.table with columns as variables. All columns must be numeric. One row per sample.}

\item{method}{string defining the type of correlation to run. The currently supported values are 'spearman', 'pearson' and 'sparcc'}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
data.frame with correlation coefficients
}
\description{
This function returns correlation coefficients for all columns in one data table against themselves.
}
