% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-correlation.R
\name{correlation}
\alias{correlation}
\title{Correlation}
\usage{
correlation(
  data1,
  data2,
  method = c("spearman", "pearson"),
  verbose = c(TRUE, FALSE),
  ...
)
}
\arguments{
\item{data1}{first dataset. An AbundanceData object or data.table}

\item{data2}{second dataset. A SampleMetadata object (if data1 is class AbundanceData) or a data.table}

\item{method}{string defining the type of correlation to run. The currently supported values are 'spearman' and 'pearson'}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
data.frame with correlation coefficients or a ComputeResult object
}
\description{
This function returns correlation coefficients for variables in one dataset against variables in a second dataset
}
