% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CorrelationComputeResult.R
\docType{class}
\name{CorrelationComputeResult-class}
\alias{CorrelationComputeResult-class}
\alias{CorrelationComputeResult}
\title{Corrleation Compute Result}
\description{
A class to represent the results of a correlation calculation. Contains ComputeResult.
This includes their representation in R, as JSON and how they are written to files.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data.frame of values where computed variables are columns and samples rows.}

\item{\code{name}}{The name of the compute, ex: 'alphaDiv'.}

\item{\code{recordIdColumn}}{The name of the column containing IDs for the samples. All other columns will be treated as computed values.}

\item{\code{ancestorIdColumns}}{A character vector of column names representing parent entities of the recordIdColumn.}

\item{\code{computedVariableMetadata}}{veupathUtils::VariableMetadataList detailing the computed variables.}

\item{\code{statistics}}{An optional slot of any values. List or data.frame are recommended. It is not required to have rows or cols map to samples.}

\item{\code{computationDetails}}{An optional message about the computed results.}

\item{\code{parameters}}{A record of the input parameters used to generate the computed results.}

\item{\code{droppedColumns}}{A character vector of column names that have been dropped for being unsuitable for the computation.}

\item{\code{data1Metadata}}{A VariableMetadata object describing the data1 object sent to the correlation function}

\item{\code{data1Variables}}{A VariableMetadataList object that describes each variable in data1.}

\item{\code{data2Metadata}}{A VariableMetadata object describing the data2 object sent to the correlation function}

\item{\code{data2Variables}}{A VariableMetadataList object that describes each variable in data2.}
}}

