% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CorrelationResult.R
\docType{class}
\name{CorrelationResult-class}
\alias{CorrelationResult-class}
\alias{CorrelationResult}
\title{Corrleation Result}
\description{
A class to represent the computed results from a correlation calculation.
This includes their representation in R, as JSON and how they are written to files.
}
\section{Slots}{

\describe{
\item{\code{statistics}}{An optional slot of any values. List or data.frame are recommended. It is not required to have rows or cols map to samples.}

\item{\code{data1Metadata}}{A character describing data1. Should at least describe the name of data1.}

\item{\code{data2Metadata}}{A character describing data2. Should at least describe the name of data2.}
}}

