% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-differentialAbundance.R
\name{differentialAbundance}
\alias{differentialAbundance}
\title{Differential abundance}
\usage{
differentialAbundance(
  data,
  comparisonVariable,
  groupA,
  groupB,
  method = c("DESeq", "ANCOMBC"),
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{data}{AbundanceData object}

\item{comparisonVariable}{string identifying the metadata column to be used when grouping samples.}

\item{groupA}{array of strings, each string matching a metadata value. Samples with this metadata value will be included in Group A}

\item{groupB}{array of strings, each string matching a metadata value. Samples with this metadata value will be included in Group B. Must be distinct from groupA.}

\item{method}{string defining the the differential abundance method. Accepted values are 'DESeq'}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
ComputeResult object
}
\description{
This function returns the fold change and associated p value for a differential abundance analysis comparing samples in two groups.
}
