% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-differentialAbundance.R
\docType{class}
\name{DifferentialAbundanceResult-class}
\alias{DifferentialAbundanceResult-class}
\alias{DifferentialAbundanceResult}
\title{Differential Abundance Result}
\description{
A DifferentialAbundanceResult object contains the results of differential
abundance analysis.
}
\section{Slots}{

\describe{
\item{\code{effectSizeLabel}}{label for the effect size metric used. Default is 'log2(Fold Change)'.}

\item{\code{statistics}}{data frame containing the statistics}

\item{\code{pValueFloor}}{the p-value floor. Any p-values less than this will be set to this value.}

\item{\code{adjustedPValueFloor}}{the adjusted p-value floor.}
}}

