% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-AbundanceData.R
\name{getAbundances}
\alias{getAbundances}
\title{Get data.table of abundances from AbundanceData}
\usage{
getAbundances(object, ignoreImputeZero = c(FALSE, TRUE))
}
\arguments{
\item{object}{AbundanceData}

\item{ignoreImputeZero}{boolean indicating whether we should respect the imputeZero slot}
}
\value{
data.table of abundances
}
\description{
Returns a data.table of abundances, respecting the
\code{imputeZero} slot.
}
