% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Comparator.R
\name{getGroupLabels}
\alias{getGroupLabels}
\title{Get labels from a group in a Comparator}
\usage{
getGroupLabels(object, group = c("groupA", "groupB"))
}
\arguments{
\item{object}{Comparator}

\item{group}{String, either "groupA" or "groupB"}
}
\value{
charactor vector of labels from groupA or groupB
}
\description{
For any Comparator object, returns the bin labels of either groupA or groupB
}
