% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-AbundanceData.R
\name{getIdColumns}
\alias{getIdColumns}
\title{Get all ID columns}
\usage{
getIdColumns(object)
}
\arguments{
\item{object}{AbundanceData}
}
\value{
vector of all ID columns
}
\description{
Returns a vector of all ID columns
}
