% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_betaDiv.R
\name{getMetadata}
\alias{getMetadata}
\title{Extract Computed Variable}
\usage{
getMetadata(computation)
}
\arguments{
\item{df}{a computation which contains an attribute called 'computedVariable'}
}
\value{
json string of Computed Variable Metadata and Details
}
\description{
This function, for any given computation,
will return Computed Variable Metadata and Details.
}
