% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomeComputations-package.R
\docType{package}
\name{microbiomeComputations-package}
\alias{microbiomeComputations}
\alias{microbiomeComputations-package}
\title{microbiomeComputations: Compute Results for Microbiome Analyses}
\description{
microbiomeComputations provides helper functions for performing common analyses on microbiome data. The results are available as json, or optionally in a `data.table`. This package is intended to be compliant with the API specified in the VEuPathDB EdaComputeService.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/VEuPathDB/microbiomeComputations}
  \item Report bugs at \url{https://github.com/VEuPathDB/microbiomeComputations/issues}
}

}
\author{
\strong{Maintainer}: Danielle Callan \email{dcallan@upenn.edu}

}
\keyword{internal}
