% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-correlation.R
\name{predicateFactory}
\alias{predicateFactory}
\title{Predicate Factory}
\usage{
predicateFactory(predicateType, threshold)
}
\arguments{
\item{predicateType}{string defining the type of predicate to run. The currently supported values are 'proportionNonZero', 'variance' and 'sd'}

\item{threshold}{numeric value associated with the predicate type}
}
\value{
Function returning a boolean indicating if a feature should be included (TRUE) or excluded (FALSE)
}
\description{
This function creates a predicate function based on a string defining the type of predicate to run and a numeric value.
The currently supported types are 'proportionNonZero', 'variance' and 'sd'. The numeric value associated
with each predicate type is the threshold for the predicate to be true.
}
