% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-AbundanceData.R
\name{pruneFeatures}
\alias{pruneFeatures}
\title{Prune features by predicate}
\usage{
pruneFeatures(object, predicate, verbose = c(TRUE, FALSE))
}
\arguments{
\item{object}{AbundanceData}

\item{predicate}{Function returning a boolean indicating if a feature should be included (TRUE) or excluded (FALSE)}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
AbundanceData with modified data slot
}
\description{
Modifies the data slot of an
AbundanceData object, to exclude features for which
the provided predicate function returns FALSE.
}
