% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rankedAbundance.R
\name{rankedAbundance}
\alias{rankedAbundance}
\title{Ranked abundance}
\usage{
rankedAbundance(
  df,
  recordIdColumn,
  method = c("median", "max", "q3", "variance"),
  cutoff = 10,
  naToZero = c(FALSE, TRUE),
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{df}{data.table with rows corresponding to records, one column indicating the record id, and all other columns corresponding to taxa abundance.}

\item{recordIdColumn}{string defining the name of the df column that specifies sample ids. Note, all other columns must be numeric and will be treated as abundance values.}

\item{method}{string defining the ranking strategy by which to order the taxa. Accepted values are 'median','max','q3',and 'variance'. Note that taxa that return a value of 0 for a given method will not be included in the results.}

\item{cutoff}{integer indicating the maximium number of taxa to be kept after ranking.}

\item{naToZero}{boolean indicating if NAs in numeric columns should be replaced with 0}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
data.table with columns recordIdColumn, followed by the top taxa as ranked by the given method, with no more taxa columns than specified by the cutoff argument.
}
\description{
This function returns abundances, ranked by a selected ranking function
}
