% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rankedAbundance.R
\name{rankedAbundanceApp}
\alias{rankedAbundanceApp}
\title{Ranked abundance app}
\usage{
rankedAbundanceApp(
  df,
  recordIdColumn,
  methods = c("median", "max", "q3", "variance"),
  cutoff = 10,
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{df}{data.frame with samples as rows, taxa as columns.}

\item{recordIdColumn}{string defining the name of the df column that specifies sample ids. Note, all other columns must be numeric and will be treated as abundance values.}

\item{methods}{vector of strings indicating ranking methods to use. Must be a subset of c('median','max','q3','variance').}

\item{cutoff}{integer indicating the maximium number of taxa to be kept after ranking.}

\item{verbose}{boolean indicating if timed logging is desired.}
}
\value{
name of a json file containing a list of data.tables, one for each method specified in methods. Each data.table contains a column for the recordIdColumn, top taxa columns, and an attribute "parameters" that records the method used.
}
\description{
This function returns the name of a json file with ranked abundance results.
}
