% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-AbundanceData.R
\name{removeIncompleteSamples}
\alias{removeIncompleteSamples}
\title{Drop samples with incomplete SampleMetadata}
\usage{
removeIncompleteSamples(object, colName = character())
}
\arguments{
\item{object}{AbundanceData}

\item{colName}{String providing the column name in SampleMetadata to check for completeness}
}
\value{
AbundanceData with modified data and sampleMetadata slots
}
\description{
Modifies the data and sampleMetadata slots of an
AbundanceData object, to exclude samples with
missing SampleMetadata for a specified column.
}
