% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-correlation.R
\name{selfCorrelation,AbundanceData-method}
\alias{selfCorrelation,AbundanceData-method}
\title{Self Correlation}
\usage{
\S4method{selfCorrelation}{AbundanceData}(
  data,
  method = c("spearman", "pearson"),
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{data}{An AbundanceData object}

\item{method}{string defining the type of correlation to run. The currently supported values are 'spearman' and 'pearson'}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
ComputeResult object
}
\description{
This function returns correlation coefficients for variables in one AbundanceData object against itself
}
