% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-correlation.R
\name{selfCorrelation,data.table-method}
\alias{selfCorrelation,data.table-method}
\title{Self Correlation}
\usage{
\S4method{selfCorrelation}{data.table}(
  data,
  method = c("spearman", "pearson"),
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{data}{a data.table}

\item{method}{string defining the type of correlation to run. The currently supported values are 'spearman' and 'pearson'}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
ComputeResult object
}
\description{
This function returns correlation coefficients for variables in one data.table against itself.
This is essentially an alias to the microbiomeComputations::correlation function.
}
