% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-correlation.R
\name{selfCorrelation}
\alias{selfCorrelation}
\title{Self Correlation}
\usage{
selfCorrelation(
  data,
  method = c("spearman", "pearson", "sparcc"),
  verbose = c(TRUE, FALSE),
  ...
)
}
\arguments{
\item{data}{first dataset. An AbundanceData object or data.table}

\item{method}{string defining the type of correlation to run. The currently supported values are 'spearman','pearson' and 'sparcc'}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
ComputeResult object
}
\description{
This function returns correlation coefficients for variables in one dataset against itself
}
