% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ComputeResult.R
\name{writeData}
\alias{writeData}
\title{Write Compute Result Data}
\usage{
writeData(object, pattern = NULL, verbose = c(TRUE, FALSE))
}
\arguments{
\item{object}{ComputeResult}

\item{pattern}{string to incorporate into tmp file name}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
filehandle where tab delimited representation of result data can be found
}
\description{
This function, for any given ComputeResult,
will return a filehandle where result data
has been written in tab delimited format.
}
