% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ComputeResult.R
\name{writeStatistics}
\alias{writeStatistics}
\title{Write Computed Variable Statistics}
\usage{
writeStatistics(object, pattern = NULL, verbose = c(TRUE, FALSE))
}
\arguments{
\item{object}{ComputeResult with statistics}

\item{pattern}{string to incorporate into tmp file name}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
filehandle where JSON representation of ComputeResult statistics can be found
}
\description{
This function, for any given ComputeResult,
will return a filehandle where the statistics table
has been written in JSON format.
}
