% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-AbundanceData.R
\docType{class}
\name{AbundanceData-class}
\alias{AbundanceData-class}
\alias{AbundanceData}
\title{Abundance Data}
\description{
A class for working with microbiome or ecological abundance data.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data.frame of abundance values with species as columns and samples as rows}

\item{\code{sampleMetadata}}{A SampleMetadata object of metadata about the samples with samples as rows and metadata variables as columns}

\item{\code{recordIdColumn}}{The name of the column containing IDs for the samples. All other columns will be treated as abundance values.}

\item{\code{ancestorIdColumns}}{A character vector of column names representing parent entities of the recordIdColumn.}

\item{\code{imputeZero}}{A logical indicating whether NA/ null values should be replaced with zeros.}

\item{\code{removeEmptySamples}}{A logical indicating whether empty (all NA/ zero)samples should be removed.}
}}

