% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_alphaDiv.R
\name{alphaDiv}
\alias{alphaDiv}
\title{Alpha diversity}
\usage{
alphaDiv(
  df,
  recordIdColumn,
  method = c("shannon", "simpson", "evenness"),
  naToZero = c(FALSE, TRUE),
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{df}{data.table with rows corresponding to records, one column indicating the record id, and all other columns corresponding to taxa abundance.}

\item{recordIdColumn}{string defining the name of the df column that specifies record ids. Note, all other columns must be numeric and will be treated as abundance values.}

\item{method}{string defining the the alpha diversity method. Accepted values are 'shannon','simpson', and 'evenness'}

\item{naToZero}{boolean indicating if NAs in numeric columns should be replaced with 0}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
data.table with columns recordIdColumn, "alphaDiversity".
}
\description{
This function returns alpha diversity values for each sample.
}
