% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_betaDiv.R
\name{betaDivApp}
\alias{betaDivApp}
\title{Beta diversity app}
\usage{
betaDivApp(
  df,
  recordIdColumn,
  methods = c("bray", "jaccard", "jsd"),
  k = 2,
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{df}{data.frame with samples as rows, taxa as columns}

\item{recordIdColumn}{string defining the name of the df column that specifies sample ids. Note, all other columns must be numeric and will be treated as abundance values.}

\item{methods}{vector of strings defining the the beta diversity dissimilarity methods to use. Must be a subset of c('bray','jaccard','jsd').}

\item{k}{integer determining the number of pcoa dimensions to return}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
name of a json file containing a list of data.tables, one for each method specified in methods. Each data.table contains a column for the recordIdColumn, as well as columns corresponding to pcoa axes.
}
\description{
This function returns the name of a json file with beta diversity results.
}
