% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-correlation.R
\name{correlation}
\alias{correlation}
\title{Correlation}
\usage{
correlation(data, method = c("spearman", "pearson"), verbose = c(TRUE, FALSE))
}
\arguments{
\item{data}{AbundanceData object}

\item{method}{string defining the type of correlation to run. The currently supported values are 'spearman' and 'pearson'}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
ComputeResult object
}
\description{
This function returns correlation coefficients for all abundance variables vs all continuous sample metadata variables.
Soon this function will be expanded to take multiple abundance datasets, and to run all-to-all correlations
for abundance variables.
}
