% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-differentialAbundance.R
\name{differentialAbundance}
\alias{differentialAbundance}
\title{Differential abundance}
\usage{
differentialAbundance(
  data,
  comparisonVariable,
  groupA,
  groupB,
  method = c("DESeq", "ANCOMBC"),
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{data}{AbsoluteAbundanceData object}

\item{comparisonVariable}{string identifying the metadata column to be used when grouping samples.}

\item{groupA}{array of strings, each string indicating a value in the comparisonVariable column. Samples with these metadata values will be included in Group A for the differential abundance calculation. Must have no overlap with groupB}

\item{groupB}{array of strings, each string indicating a value in the comparisonVariable column. Samples with these metadata values will be included in Group B for the differential abundance calculation. Must have no overlap with groupA.}

\item{method}{string defining the the differential abundance method. Accepted values are 'DESeq' and 'ANCOMBC'.}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
ComputeResult object
}
\description{
This function returns the fold change and associated p value for a differential abundance analysis comparing samples in two groups.
}
