% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-AbundanceData.R
\name{getAbundances}
\alias{getAbundances}
\alias{getAbundances,AbundanceData-method}
\title{Get data.table of abundances from AbundanceData}
\usage{
getAbundances(
  object,
  ignoreImputeZero = c(FALSE, TRUE),
  includeIds = c(TRUE, FALSE),
  verbose = c(TRUE, FALSE)
)

\S4method{getAbundances}{AbundanceData}(
  object,
  ignoreImputeZero = c(FALSE, TRUE),
  includeIds = c(TRUE, FALSE),
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{object}{AbundanceData}

\item{ignoreImputeZero}{boolean indicating whether we should respect the imputeZero slot}

\item{includeIds}{boolean indicating whether we should include recordIdColumn and ancestorIdColumns}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
data.table of abundances
}
\description{
Returns a data.table of abundances, respecting the
\code{imputeZero} slot.
}
