% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_abundance.R
\name{rankedAbundance}
\alias{rankedAbundance}
\title{Ranked abundance}
\usage{
rankedAbundance(
  otu,
  method = c("median", "max", "q3", "var"),
  cutoff = 10,
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{otu}{data.frame with samples as rows, taxa as columns}

\item{method}{string defining the ranking strategy by which to order the taxa. Accepted values are 'median','max','q3',and 'var'.}

\item{cutoff}{integer indicating the maximium number of taxa to be kept after ranking.}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
something that's useful. TBD
}
\description{
This function returns abundances, ranked by a selected ranking function
}
