% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_abundance.R
\name{rankedAbundanceApp}
\alias{rankedAbundanceApp}
\title{Ranked abundance app}
\usage{
rankedAbundanceApp(otu, cutoff = 10, verbose = c(TRUE, FALSE))
}
\arguments{
\item{otu}{data.frame with samples as rows, taxa as columns}

\item{cutoff}{integer indicating the maximium number of taxa to be kept after ranking.}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
something that's useful. TBD
}
\description{
This function returns the name of a json file with ranked abundance results.
}
