% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-AbundanceData.R
\name{removeIncompleteSamples}
\alias{removeIncompleteSamples}
\alias{removeIncompleteSamples,AbundanceData-method}
\title{Drop samples with incomplete SampleMetadata}
\usage{
removeIncompleteSamples(
  object,
  colName = character(),
  verbose = c(TRUE, FALSE)
)

\S4method{removeIncompleteSamples}{AbundanceData}(
  object,
  colName = character(),
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{object}{AbundanceData}

\item{colName}{String providing the column name in SampleMetadata to check for completeness}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
AbundanceData with modified data and sampleMetadata slots
}
\description{
Modifies the data and sampleMetadata slots of an
AbundanceData object, to exclude samples with
missing SampleMetadata for a specified column.
}
