% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-correlation.R
\name{selfCorrelation,AbundanceData-method}
\alias{selfCorrelation,AbundanceData-method}
\title{Self Correlation}
\usage{
\S4method{selfCorrelation}{AbundanceData}(
  data,
  method = c("spearman", "pearson"),
  verbose = c(TRUE, FALSE),
  proportionNonZeroThreshold = 0.5,
  varianceThreshold = 0,
  stdDevThreshold = 0
)
}
\arguments{
\item{data}{An AbundanceData object}

\item{method}{string defining the type of correlation to run. The currently supported values are 'spearman' and 'pearson'}

\item{verbose}{boolean indicating if timed logging is desired}

\item{proportionNonZeroThreshold}{numeric threshold to filter features by proportion of non-zero values across samples}

\item{varianceThreshold}{numeric threshold to filter features by variance across samples}

\item{stdDevThreshold}{numeric threshold to filter features by standard deviation across samples}
}
\value{
ComputeResult object
}
\description{
This function returns correlation coefficients for variables in one AbundanceData object against itself
}
