% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ComputeResult.R
\name{writeData}
\alias{writeData}
\title{Write Compute Result Data}
\usage{
writeData(object, pattern = NULL, verbose = c(TRUE, FALSE))
}
\arguments{
\item{object}{ComputeResult}
}
\value{
filehandle where tab delimited representation of result data can be found
}
\description{
This function, for any given ComuteResult,
will return a filehandle where result data
has been written in tab delimited format.
}
