% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Compute.R
\name{writeMeta}
\alias{writeMeta}
\title{Write Computed Variable Metadata}
\usage{
writeMeta(object, pattern = NULL, verbose = c(TRUE, FALSE))
}
\arguments{
\item{object}{ComputeResult}
}
\value{
filehandle where JSON representation of ComputedVariableMetadata can be found
}
\description{
This function, for any given ComuteResult,
will return a filehandle where ComputedVariableMetadata
has been written in JSON format.
}
